Adding all

+++ user/user.h
@@ -26,9 +26,9 @@ int sleep(int);
 int uptime(void);
 int getprocs(struct pstat*);
 //Adding declaration
-int freepmem(void);
+uint64 freepmem(void);
 int memoryuser(void);
-void* mmap(void*,uint64,int,int,int,int);
+void *mmap(void*,uint64,int,int,int,int);
 int munmap(void*,uint64);
 
 // ulib.c
